chore: TDET-50 File folder structure redone
